plugins: fix meson.build
We create gst project from:

 - gst-project-maker
 - gst-element-maker

For this we need to fix the meson.build file base on our needs. To be
clear:

Since it is an element created by inheriting videofilter class, it is
necessary to add a dependency to meson.build.
Other thing is that function gst_element_register is duplicated, so
delete simplefilterplugin.c file and update plugin meson.build

Signed-off-by: Tommaso Merciai <tomm.merciai@gmail.com>